se crea JWT

+++ user-service/src/main/java/com/ecomarket/user_service/service/UsuarioService.java
@@ -19,7 +19,7 @@ public class UsuarioService {
         return usuarioRepository.findAll();
     }
 
-    public Usuario findById(Long id) {
+    public Usuario findById(Integer id) {
         return usuarioRepository.findById(id).orElse(null);
     }
 
@@ -27,7 +27,7 @@ public class UsuarioService {
         return usuarioRepository.save(usuario);
     }
 
-    public void delete(Long id) {
+    public void delete(Integer id) {
         usuarioRepository.deleteById(id);
     }
 }
